feat: update the animation to ready button
